A2/requirements.txt
torch
transformers
datasets
Pillow
tqdm
scikit-learn
pytest
openai
coverage